Rename existing add, edit, delete commands parsers to become applicants
