+++ src/test/java/seedu/address/logic/parser/DeleteApplicantCommandParserTest.java
@@ -8,6 +8,7 @@ import static seedu.address.testutil.TypicalIndexes.INDEX_FIRST_PERSON;
 import org.junit.jupiter.api.Test;
 
 import seedu.address.logic.commands.DeleteCommand;
+import seedu.address.logic.parser.applicants.DeleteApplicantCommandParser;
 
 /**
  * As we are only doing white-box testing, our test cases do not cover path variations
@@ -16,9 +17,9 @@ import seedu.address.logic.commands.DeleteCommand;
  * The path variation for those two cases occur inside the ParserUtil, and
  * therefore should be covered by the ParserUtilTest.
  */
-public class DeleteCommandParserTest {
+public class DeleteApplicantCommandParserTest {
 
-    private DeleteCommandParser parser = new DeleteCommandParser();
+    private DeleteApplicantCommandParser parser = new DeleteApplicantCommandParser();
 
     @Test
     public void parse_validArgs_returnsDeleteCommand() {
